+++ modules/coralogix-aws-shipper/Sns.tf
@@ -24,4 +24,5 @@ resource "aws_sns_topic_subscription" "lambda_sns_subscription" {
   protocol      = "lambda"
   endpoint      = module.lambda.integration.lambda_function_arn
   filter_policy = var.sns_topic_filter != null ? jsonencode(var.sns_topic_filter) : null
+  filter_policy_scope = var.sns_topic_filter_policy_scope != null ? var.sns_topic_filter_policy_scope : null
 }